+++ lib/models/transaction_info.dart
@@ -1,12 +1,12 @@
 class TransactionInfo {
   final String payerPhoneNumber, referenceId, invoiceId, currency, description;
-  final double amount;
+  final num amount;
 
   TransactionInfo({
     required this.payerPhoneNumber,
     required this.amount,
     this.referenceId = '',
-    this.invoiceId = '',
+    required this.invoiceId,
     this.currency = 'USD',
     this.description = 'Payment went through.',
   });