+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
-
 Hormuud EVCPLUS Merchant API Integration
+
+## 0.0.2
+Merchant URL Change Applied
+InvoiceId Required
+Example App Changes